refactor(audit): extract render and command dispatch from AuditDocument
Moves all render_* methods and display helpers to audit/render.rs, and
all command dispatch, keyboard navigation, and context menu logic to
audit/commands.rs, following the same impl AuditDocument split pattern
used in data_grid_panel/render.rs.

Also migrates AuditDocument to emit DocumentEvent directly (T-5-2) and
updates DocumentHandle::Audit subscription to forward it without
translation (T-5-3).

audit/mod.rs: 3454 → 1611 LOC (H2 gate: below 1727 ✓)
